requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==6.7
Flask==1.1.2
fritzconnection==1.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libnacl==1.7.2
Logbook==1.5.3
MarkupSafe==1.1.1
memoization==0.2.3
multidict==5.1.0
requests==2.25.1
threema.gateway==4.0.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
